Split NirCodeGen into smaller modules
