+++ nscplugin/src/main/scala/scala/scalanative/nscplugin/NirGenExpr.scala
@@ -1,25 +1,14 @@
 package scala.scalanative
 package nscplugin
 
-import java.nio.file.Path
-import scala.collection.mutable, mutable.UnrolledBuffer
-import scala.tools.nsc.{util => _, _}
-import scala.tools.nsc.plugins._
-import scala.util.{Either, Left, Right}
+import scala.collection.mutable
 import scala.reflect.internal.Flags._
-import util._, util.ScopedVar.scoped
-import nir.{InstBuffer => Buf, _}
-import NirPrimitives._
+import scalanative.nir._
+import scalanative.util.unsupported
+import scalanative.util.ScopedVar.scoped
+import scalanative.nscplugin.NirPrimitives._
 
-abstract class NirCodeGen
-    extends PluginComponent
-    with NirFiles
-    with NirTypeEncoding
-    with NirNameEncoding {
-  val nirAddons: NirGlobalAddons {
-    val global: NirCodeGen.this.global.type
-  }
-
+trait NirGenExpr { self: NirGenPhase =>
   import global._
   import definitions._
   import treeInfo.hasSynthCaseSymbol
@@ -27,446 +16,8 @@ abstract class NirCodeGen
   import nirDefinitions._
   import SimpleType.{fromType, fromSymbol}
 
-  val phaseName = "nir"
-
-  private val curLazyAnonDefs =
-    new util.ScopedVar[mutable.Map[Symbol, ClassDef]]
-  private val curClassSym   = new util.ScopedVar[Symbol]
-  private val curMethodSym  = new util.ScopedVar[Symbol]
-  private val curMethodInfo = new util.ScopedVar[CollectMethodInfo]
-  private val curMethodEnv  = new util.ScopedVar[MethodEnv]
-  private val curMethodThis = new util.ScopedVar[Option[Val]]
-  private val curFresh      = new util.ScopedVar[nir.Fresh]
-  private val curUnwind     = new util.ScopedVar[nir.Next]
-
-  private def lazyAnonDefs =
-    curLazyAnonDefs.get
-  private def consumeLazyAnonDef(sym: Symbol): ClassDef = {
-    lazyAnonDefs
-      .get(sym)
-      .fold {
-        sys.error(s"Couldn't find anon def for $sym")
-      } { cd =>
-        lazyAnonDefs.remove(cd.symbol)
-        cd
-      }
-  }
-
-  override def newPhase(prev: Phase): StdPhase =
-    new NirCodePhase(prev)
-
   final case class ValTree(value: nir.Val)    extends Tree
   final case class ContTree(f: () => nir.Val) extends Tree
-
-  class NirCodePhase(prev: Phase) extends StdPhase(prev) {
-    override def run(): Unit = {
-      scalaPrimitives.init()
-      nirPrimitives.init()
-      super.run()
-    }
-
-    override def apply(cunit: CompilationUnit): Unit = {
-      val classDefs    = mutable.UnrolledBuffer.empty[ClassDef]
-      val lazyAnonDefs = mutable.Map.empty[Symbol, ClassDef]
-
-      def collectClassDefs(tree: Tree): Unit = tree match {
-        case EmptyTree =>
-          ()
-        case PackageDef(_, stats) =>
-          stats.foreach(collectClassDefs)
-        case cd: ClassDef =>
-          val sym = cd.symbol
-          if (sym.isAnonymousFunction) {
-            lazyAnonDefs(sym) = cd
-          } else if (isPrimitiveValueClass(sym) || (sym == ArrayClass)) {
-            ()
-          } else {
-            classDefs += cd
-          }
-      }
-
-      def genClass(cd: ClassDef): Unit = {
-        val path   = genPathFor(cunit, cd.symbol)
-        val buffer = new StatBuffer
-        buffer.genClass(cd)
-        //println(s"--- " + path)
-        //buffer.toSeq.foreach(defn => println(defn.show))
-        genIRFile(path, buffer.toSeq)
-      }
-
-      scoped(
-        curLazyAnonDefs := lazyAnonDefs
-      ) {
-        collectClassDefs(cunit.body)
-        classDefs.foreach(genClass)
-        lazyAnonDefs.values.foreach(genClass)
-      }
-    }
-  }
-
-  class MethodEnv(val fresh: Fresh) {
-    private val env = mutable.Map.empty[Symbol, Val]
-
-    def enter(sym: Symbol, value: Val): Unit = {
-      env += ((sym, value))
-    }
-
-    def enterLabel(ld: LabelDef): Local = {
-      val local = fresh()
-      enter(ld.symbol, Val.Local(local, Type.Ptr))
-      local
-    }
-
-    def resolve(sym: Symbol): Val = {
-      env(sym)
-    }
-
-    def resolveLabel(ld: LabelDef): Local = {
-      val Val.Local(n, Type.Ptr) = resolve(ld.symbol)
-      n
-    }
-  }
-
-  class CollectMethodInfo extends Traverser {
-    var mutableVars = Set.empty[Symbol]
-    var labels      = Set.empty[LabelDef]
-
-    override def traverse(tree: Tree) = {
-      tree match {
-        case label: LabelDef =>
-          labels += label
-        case Assign(id @ Ident(_), _) =>
-          mutableVars += id.symbol
-        case _ =>
-          ()
-      }
-      super.traverse(tree)
-    }
-
-    def collect(tree: Tree) = {
-      traverse(tree)
-      this
-    }
-  }
-
-  class StatBuffer extends nir.DefnBuffer { buf =>
-    def genClass(cd: ClassDef): Unit = {
-      scoped(
-        curClassSym := cd.symbol
-      ) {
-        if (cd.symbol.isStruct) genStruct(cd)
-        else genNormalClass(cd)
-      }
-    }
-
-    def genStruct(cd: ClassDef): Unit = {
-      val sym    = cd.symbol
-      val attrs  = genStructAttrs(sym)
-      val name   = genTypeName(sym)
-      val fields = genStructFields(sym)
-      val body   = cd.impl.body
-
-      buf += Defn.Struct(attrs, name, fields)
-      genMethods(cd)
-    }
-
-    def genStructAttrs(sym: Symbol): Attrs = Attrs.None
-
-    def genNormalClass(cd: ClassDef): Unit = {
-      val sym    = cd.symbol
-      def attrs  = genClassAttrs(cd)
-      def name   = genTypeName(sym)
-      def parent = genClassParent(sym)
-      def traits = genClassInterfaces(sym)
-
-      genClassFields(sym)
-      genMethods(cd)
-
-      buf += {
-        if (sym.isScalaModule) Defn.Module(attrs, name, parent, traits)
-        else if (sym.isInterface) Defn.Trait(attrs, name, traits)
-        else Defn.Class(attrs, name, parent, traits)
-      }
-    }
-
-    def genClassParent(sym: Symbol): Option[nir.Global] =
-      if (sym == NObjectClass) None
-      else if (sym.superClass == NoSymbol || sym.superClass == ObjectClass)
-        Some(genTypeName(NObjectClass))
-      else Some(genTypeName(sym.superClass))
-
-    def genClassAttrs(cd: ClassDef): Attrs = {
-      val sym = cd.symbol
-      def pinned = {
-        val ctor =
-          if (!sym.isScalaModule || sym.isExternModule) {
-            Seq()
-          } else {
-            Seq(Attr.PinAlways(genMethodName(sym.asClass.primaryConstructor)))
-          }
-        val annotated = for {
-          decl <- sym.info.decls
-          if decl.hasAnnotation(PinClass)
-        } yield {
-          Attr.PinAlways(genName(decl))
-        }
-
-        annotated.toSeq ++ ctor
-      }
-      val attrs = sym.annotations.collect {
-        case ann if ann.symbol == ExternClass =>
-          Attr.Extern
-        case ann if ann.symbol == LinkClass =>
-          val Apply(_, Seq(Literal(Constant(name: String)))) = ann.tree
-          Attr.Link(name)
-      }
-      val pure = if (PureModules.contains(sym)) Seq(Attr.Pure) else Seq()
-
-      val weak = cd.impl.body.collect {
-        case dd: DefDef =>
-          Attr.PinWeak(genMethodName(dd.symbol))
-      }
-
-      Attrs.fromSeq(pinned ++ pure ++ attrs ++ weak)
-    }
-
-    def genClassInterfaces(sym: Symbol) =
-      for {
-        parent <- sym.info.parents
-        psym = parent.typeSymbol if psym.isInterface
-      } yield {
-        genTypeName(psym)
-      }
-
-    def genClassFields(sym: Symbol): Unit = {
-      val attrs = nir.Attrs(isExtern = sym.isExternModule)
-
-      for (f <- sym.info.decls if f.isField) {
-        val ty   = genType(f.tpe, box = false)
-        val name = genFieldName(f)
-
-        buf += Defn.Var(attrs, name, ty, Val.None)
-      }
-    }
-
-    def genMethods(cd: ClassDef): Unit =
-      cd.impl.body.foreach {
-        case dd: DefDef =>
-          genMethod(dd)
-        case _ =>
-          ()
-      }
-
-    def genMethod(dd: DefDef): Unit = {
-      val fresh = new Fresh("src")
-      val env   = new MethodEnv(fresh)
-
-      scoped(
-        curMethodSym := dd.symbol,
-        curMethodEnv := env,
-        curMethodInfo := (new CollectMethodInfo).collect(dd.rhs),
-        curFresh := fresh,
-        curUnwind := Next.None
-      ) {
-        val sym      = dd.symbol
-        val owner    = curClassSym.get
-        val attrs    = genMethodAttrs(sym)
-        val name     = genMethodName(sym)
-        val isStatic = owner.isExternModule || owner.isImplClass
-        val sig      = genMethodSig(sym, isStatic)
-        val params   = genParams(dd, isStatic)
-
-        dd.rhs match {
-          case EmptyTree =>
-            buf += Defn.Declare(attrs, name, sig)
-
-          case _ if dd.name == nme.CONSTRUCTOR && owner.isExternModule =>
-            validateExternCtor(dd.rhs)
-            ()
-
-          case _ if dd.name == nme.CONSTRUCTOR && owner.isStruct =>
-            ()
-
-          case rhs if owner.isExternModule =>
-            genExternMethod(attrs, name, sig, params, rhs)
-
-          case rhs =>
-            val body = genNormalMethodBody(dd, params, rhs, isStatic)
-            buf += Defn.Define(attrs, name, sig, body)
-        }
-      }
-    }
-
-    def genExternMethod(attrs: nir.Attrs,
-                        name: nir.Global,
-                        sig: nir.Type,
-                        params: Seq[nir.Val.Local],
-                        rhs: Tree): Unit = {
-      rhs match {
-        case Apply(ref: RefTree, Seq()) if ref.symbol == ExternMethod =>
-          val moduleName  = genTypeName(curClassSym)
-          val externAttrs = Attrs(isExtern = true)
-          val externDefn  = Defn.Declare(externAttrs, name, sig)
-
-          buf += externDefn
-
-        case _ if curMethodSym.hasFlag(ACCESSOR) =>
-          ()
-
-        case rhs =>
-          unsupported("methods in extern objects must have extern body")
-      }
-    }
-
-    def validateExternCtor(rhs: Tree): Unit = {
-      val Block(_ +: init, _) = rhs
-      val externs = init.map {
-        case Assign(ref: RefTree, Apply(extern, Seq()))
-            if extern.symbol == ExternMethod =>
-          ref.symbol
-        case _ =>
-          unsupported(
-            "extern objects may only contain " + "extern fields and methods")
-      }.toSet
-      for {
-        f <- curClassSym.info.decls if f.isField
-        if !externs.contains(f)
-      } {
-        unsupported("extern objects may only contain extern fields")
-      }
-    }
-
-    def genMethodAttrs(sym: Symbol): Attrs = {
-      val inlineAttrs = {
-        if (sym.hasFlag(ACCESSOR)) {
-          Seq(Attr.AlwaysInline)
-        } else {
-          sym.annotations.collect {
-            case ann if ann.symbol == NoInlineClass   => Attr.NoInline
-            case ann if ann.symbol == InlineHintClass => Attr.InlineHint
-            case ann if ann.symbol == InlineClass     => Attr.AlwaysInline
-          }
-        }
-      }
-      val overrideAttrs: Seq[Attr] = {
-        val owner = sym.owner
-        if (owner.primaryConstructor eq sym) {
-          genConstructorOverridePins(owner)
-        } else {
-          sym.overrides.collect {
-            case ovsym if !sym.owner.asClass.isTrait =>
-              Attr.Override(genMethodName(ovsym))
-          }
-        }
-      }
-      val pureAttrs = {
-        if (PureMethods.contains(sym)) Seq(Attr.Pure) else Seq()
-      }
-
-      Attrs.fromSeq(inlineAttrs ++ overrideAttrs ++ pureAttrs)
-    }
-
-    def genConstructorOverridePins(owner: Symbol): Seq[nir.Attr] = {
-      val traits = owner.info.baseClasses.map(_.asClass).filter(_.isTrait)
-      val traitMethods =
-        mutable.Map.empty[String, mutable.UnrolledBuffer[nir.Global]]
-
-      traits.foreach { trt =>
-        trt.info.declarations.foreach { meth =>
-          val name = genName(meth)
-          val sig  = name.id
-          if (!traitMethods.contains(sig)) {
-            traitMethods(sig) = mutable.UnrolledBuffer.empty[nir.Global]
-          }
-          traitMethods(sig) += name
-        }
-      }
-
-      val pins = mutable.UnrolledBuffer.empty[nir.Attr]
-
-      owner.info.declarations.foreach { decl =>
-        decl.overrides.foreach { ovsym =>
-          if (!ovsym.owner.asClass.isTrait) {
-            pins += Attr.PinIf(genMethodName(decl), genMethodName(ovsym))
-          }
-        }
-      }
-
-      owner.info.members.foreach { decl =>
-        if (decl.isMethod && decl.owner != ObjectClass) {
-          val declname = genName(decl)
-          val sig      = declname.id
-          val methods  = traitMethods.get(sig).getOrElse(Seq.empty)
-          methods.foreach { methname =>
-            pins += Attr.PinIf(declname, methname)
-          }
-        }
-      }
-
-      pins
-    }
-
-    def genParams(dd: DefDef, isStatic: Boolean): Seq[Val.Local] =
-      genParamSyms(dd, isStatic).map {
-        case None =>
-          val fresh = curFresh.get
-          Val.Local(fresh(), genType(curClassSym.tpe, box = true))
-        case Some(sym) =>
-          val fresh = curFresh.get
-          val name  = fresh()
-          val ty    = genType(sym.tpe, box = false)
-          val param = Val.Local(name, ty)
-          curMethodEnv.enter(sym, param)
-          param
-      }
-
-    def genNormalMethodBody(dd: DefDef,
-                            params: Seq[Val.Local],
-                            bodyp: Tree,
-                            isStatic: Boolean): Seq[nir.Inst] = {
-      val fresh = curFresh.get
-      val buf   = new ExprBuffer()(fresh)
-
-      def genPrelude(): Unit = {
-        val vars = curMethodInfo.mutableVars.toSeq
-        buf.label(fresh(), params)
-        vars.foreach { sym =>
-          val ty    = genType(sym.info, box = false)
-          val alloc = buf.stackalloc(ty, Val.None)
-          curMethodEnv.enter(sym, alloc)
-        }
-      }
-
-      def genBody(): Val = bodyp match {
-        // Tailrec emits magical labeldefs that can hijack this reference is
-        // current method. This requires special treatment on our side.
-        case Block(List(ValDef(_, nme.THIS, _, _)),
-                   label @ LabelDef(name, Ident(nme.THIS) :: _, rhs)) =>
-          val local  = curMethodEnv.enterLabel(label)
-          val values = params.take(label.params.length)
-
-          buf.jump(local, values)
-          buf.genTailRecLabel(dd, isStatic, label)
-
-        case _ if curMethodSym.get == NObjectInitMethod =>
-          nir.Val.Unit
-
-        case _ =>
-          scoped(
-            curMethodThis := {
-              if (isStatic) None
-              else Some(Val.Local(params.head.name, params.head.ty))
-            }
-          ) {
-            buf.genExpr(bodyp)
-          }
-      }
-
-      genPrelude()
-      buf.ret(genBody())
-      buf.toSeq
-    }
-  }
 
   class FixupInstBuffer(implicit fresh: Fresh) extends nir.InstBuffer {
     private var labeled = false
@@ -497,7 +48,6 @@ abstract class NirCodeGen
   }
 
   class ExprBuffer(implicit fresh: Fresh) extends FixupInstBuffer { buf =>
-
     def genExpr(tree: Tree): Val = tree match {
       case EmptyTree =>
         Val.Unit
@@ -2137,159 +1687,5 @@ abstract class NirCodeGen
           None
       }
     }
-  }
-
-  def genParamSyms(dd: DefDef, isStatic: Boolean): Seq[Option[Symbol]] = {
-    val vp     = dd.vparamss
-    val params = if (vp.isEmpty) Nil else vp.head.map(p => Some(p.symbol))
-    if (isStatic) params else None +: params
-  }
-
-  def unwrapClassTagOption(tree: Tree): Option[Symbol] =
-    tree match {
-      case Typed(Apply(ref: RefTree, args), _) =>
-        ref.symbol match {
-          case ByteClassTag    => Some(ByteClass)
-          case ShortClassTag   => Some(ShortClass)
-          case CharClassTag    => Some(CharClass)
-          case IntClassTag     => Some(IntClass)
-          case LongClassTag    => Some(LongClass)
-          case FloatClassTag   => Some(FloatClass)
-          case DoubleClassTag  => Some(DoubleClass)
-          case BooleanClassTag => Some(BooleanClass)
-          case UnitClassTag    => Some(UnitClass)
-          case AnyClassTag     => Some(AnyClass)
-          case ObjectClassTag  => Some(ObjectClass)
-          case AnyValClassTag  => Some(ObjectClass)
-          case AnyRefClassTag  => Some(ObjectClass)
-          case NothingClassTag => Some(NothingClass)
-          case NullClassTag    => Some(NullClass)
-          case ClassTagApply =>
-            val Seq(Literal(const: Constant)) = args
-            Some(const.typeValue.typeSymbol)
-          case _ =>
-            None
-        }
-
-      case tree =>
-        None
-    }
-
-  def unwrapTagOption(tree: Tree): Option[SimpleType] = {
-    tree match {
-      case Apply(ref: RefTree, args) =>
-        def allsts = {
-          val sts = args.flatMap(unwrapTagOption(_).toSeq)
-          if (sts.length == args.length) Some(sts) else None
-        }
-        def just(sym: Symbol) = Some(SimpleType(sym))
-        def wrap(sym: Symbol) = allsts.map(SimpleType(sym, _))
-
-        ref.symbol match {
-          case UnitTagMethod    => just(UnitClass)
-          case BooleanTagMethod => just(BooleanClass)
-          case CharTagMethod    => just(CharClass)
-          case ByteTagMethod    => just(ByteClass)
-          case UByteTagMethod   => just(UByteClass)
-          case ShortTagMethod   => just(ShortClass)
-          case UShortTagMethod  => just(UShortClass)
-          case IntTagMethod     => just(IntClass)
-          case UIntTagMethod    => just(UIntClass)
-          case LongTagMethod    => just(LongClass)
-          case ULongTagMethod   => just(ULongClass)
-          case FloatTagMethod   => just(FloatClass)
-          case DoubleTagMethod  => just(DoubleClass)
-          case PtrTagMethod     => just(PtrClass)
-          case RefTagMethod     => just(unwrapClassTagOption(args.head).get)
-          case sym if NatBaseTagMethod.contains(sym) =>
-            just(NatBaseClass(NatBaseTagMethod.indexOf(sym)))
-          case NatDigitTagMethod =>
-            wrap(NatDigitClass)
-          case CArrayTagMethod =>
-            wrap(CArrayClass)
-          case sym if CStructTagMethod.contains(sym) =>
-            wrap(CStructClass(args.length))
-          case _ =>
-            None
-        }
-
-      case tree =>
-        None
-    }
-  }
-
-  def unwrapTag(tree: Tree): SimpleType =
-    unwrapTagOption(tree).getOrElse {
-      unsupported(s"can't recover runtime tag from $tree")
-    }
-
-  object CVararg {
-    def unapply(tree: Tree): Option[(SimpleType, Tree)] = tree match {
-      case Apply(fun, Seq(argp, tagp)) if fun.symbol == CVarargMethod =>
-        val st = unwrapTag(tagp)
-        Some((st, argp))
-      case _ =>
-        None
-    }
-  }
-
-  object MaybeAsInstanceOf {
-    def unapply(tree: Tree): Some[Tree] = tree match {
-      case Apply(TypeApply(asInstanceOf_? @ Select(base, _), _), _)
-          if asInstanceOf_?.symbol == Object_asInstanceOf =>
-        Some(base)
-      case _ =>
-        Some(tree)
-    }
-  }
-
-  object MaybeBlock {
-    def unapply(tree: Tree): Some[Tree] = tree match {
-      case Block(Seq(), MaybeBlock(expr)) => Some(expr)
-      case _                              => Some(tree)
-    }
-  }
-
-  object WrapArray {
-    lazy val isWrapArray: Set[Symbol] =
-      Seq(
-        nme.wrapRefArray,
-        nme.wrapByteArray,
-        nme.wrapShortArray,
-        nme.wrapCharArray,
-        nme.wrapIntArray,
-        nme.wrapLongArray,
-        nme.wrapFloatArray,
-        nme.wrapDoubleArray,
-        nme.wrapBooleanArray,
-        nme.wrapUnitArray,
-        nme.genericWrapArray
-      ).map(getMemberMethod(PredefModule, _)).toSet
-
-    def unapply(tree: Apply): Option[Tree] = tree match {
-      case Apply(wrapArray_?, List(wrapped))
-          if isWrapArray(wrapArray_?.symbol) =>
-        Some(wrapped)
-      case _ =>
-        None
-    }
-  }
-
-  object ExternForwarder {
-    // format: OFF
-    def unapply(tree: Tree): Option[Symbol] = tree match {
-      case DefDef(_, _, _, Seq(params), _, Apply(sel @ Select(from, _), args))
-          if from.symbol != null
-          && from.symbol.isExternModule
-          && params.length == args.length
-          && params.zip(args).forall {
-               case (param, arg: RefTree) => param.symbol == arg.symbol
-               case _                     => false
-             } =>
-        Some(sel.symbol)
-      case _ =>
-        None
-    }
-    // format: ON
   }
 }
