+++ nscplugin/src/main/scala/scala/scalanative/nscplugin/NirPlugin.scala
@@ -7,16 +7,16 @@ import scala.tools.nsc.plugins._
 class NirPlugin(val global: Global) extends Plugin { self =>
   val name        = "nir"
   val description = "Compile to Scala Native IR (NIR)"
-  val components  = List[PluginComponent](nirCodeGen)
+  val components  = List[PluginComponent](nirGen)
 
   object nirAddons extends {
     val global: NirPlugin.this.global.type = NirPlugin.this.global
   } with NirGlobalAddons
 
-  object nirCodeGen extends {
+  object nirGen extends {
     val global: self.global.type                 = self.global
     val nirAddons: NirPlugin.this.nirAddons.type = NirPlugin.this.nirAddons
     override val runsAfter                       = List("mixin")
     override val runsBefore                      = List("delambdafy", "cleanup", "terminal")
-  } with NirCodeGen
+  } with NirGenPhase
 }
